Refactor preparation-reunions page and fix errors
Major Changes:
- Reduced preparation_reunions.js from 1224 to 416 lines (66% reduction)
- Created modular architecture with 4 specialized modules:
  * ActivityManager: Handles activity row operations and rendering
  * FormManager: Handles form population and data extraction
  * DateManager: Handles date navigation and calculations
  * PrintManager: Handles printing functionality

Bug Fixes:
- Fixed 'this.activities.find is not a function' error by properly handling
  API response formats (array vs object with data property)
- Fixed duplicate key constraint violation in save-reunion-preparation
  endpoint by using atomic UPSERT (INSERT ... ON CONFLICT DO UPDATE)
  instead of separate check-then-insert/update operations

Benefits:
- Much more maintainable and testable code structure
- Elimination of race conditions in database operations
- Clearer separation of concerns
- Easier to extend and debug
